+++ tests/test_indices.py
@@ -131,6 +131,24 @@ class TestColdSpellDays:
         assert out.units == "days"
 
 
+class TestColdSpellFreq:
+    def test_simple(self, tas_series):
+        a = np.zeros(365)
+        a[10:20] -= 15  # 10 days
+        a[40:43] -= 50  # too short -> 0
+        a[80:86] -= 30
+        a[95:101] -= 30
+        da = tas_series(a + K2C, start="1971-01-01")
+
+        out = xci.cold_spell_frequency(da, thresh="-10. C", freq="M")
+        np.testing.assert_array_equal(out, [1, 0, 1, 1, 0, 0, 0, 0, 0, 0, 0, 0])
+        assert out.units == ""
+
+        out = xci.cold_spell_frequency(da, thresh="-10. C", freq="YS")
+        np.testing.assert_array_equal(out, 3)
+        assert out.units == ""
+
+
 class TestConsecutiveFrostDays:
     def test_one_freeze_day(self, tasmin_series):
         a = tasmin_series(np.array([3, 4, 5, -1, 3]) + K2C)
@@ -899,6 +917,30 @@ class TestTgMaxTgMinIndices:
     #
     #     np.testing.assert_array_less(-dtr, [0, 0])
     #     np.testing.assert_allclose([dtr.mean()], [20], atol=10)
+    @pytest.mark.parametrize(
+        "op,expected",
+        [
+            ("max", 12.5),
+            (np.max, 12.5),
+            ("min", 4.0),
+            (np.min, 4.0),
+            ("std", 2.72913233),
+            (np.std, 2.72913233),
+        ],
+    )
+    def test_static_reduce_daily_temperature_range(
+        self, tasmin_series, tasmax_series, op, expected
+    ):
+        tasmin, tasmax = self.static_tmin_tmax_setup(tasmin_series, tasmax_series)
+        dtr = xci.daily_temperature_range(tasmin, tasmax, freq="YS", op=op)
+        assert dtr.units == "K"
+
+        if isinstance(op, str):
+            output = getattr(np, op)(tasmax - tasmin)
+        else:
+            output = op(tasmax - tasmin)
+        np.testing.assert_array_almost_equal(dtr, expected)
+        np.testing.assert_equal(dtr, output)
 
     def test_static_daily_temperature_range(self, tasmin_series, tasmax_series):
         tasmin, tasmax = self.static_tmin_tmax_setup(tasmin_series, tasmax_series)